[#feature_164] refactor: EmitterRepository 인터페이스 생성
